requirements.txt
requests>=2.31.0
prometheus-remote-writer>=0.1.3
python-dotenv>=1.0.0
redis>=7.2.0
geopy>=2.4.1